requirements.txt
Flask==2.3.2
requests==2.31.0
openai==1.30.1
gtts==2.5.4
playsound==1.2.2
python-dotenv==1.0.1
gunicorn==21.2.0
